장바구니, 주문 css 작업

+++ src/main/java/com/nike/web/domain/CartDTO.java
@@ -24,7 +24,7 @@ public class CartDTO {
 	private String proName;
 	private int productSize;
 	private int proPrice;
-	private long proDiscount;
+	private double proDiscount;
 
 	//	Product Image
 	private int proimgNo;
